+++ src/commands/clear-all-categories.ts
@@ -4,7 +4,8 @@ import BookmarkManager from '../services/bookmark-manager';
 import { inject, injectable} from 'inversify';
 import { CLEAR_ALL_CATEGORIES } from '../consts/commands';
 import { CATEGORY_CONFIRM_CLEAR_ALL, 
-	GENERIC_YES_ANSWER, GENERIC_NO_ANSWER } from '../consts/messages';
+				 GENERIC_YES_ANSWER, 
+				 GENERIC_NO_ANSWER } from '../consts/constants-messages';
 import { Command } from './base/command';
 import { DataStorageManager } from '../services/data-storage-manager';
 import { FAVORITE_REPOS_KEY } from '../consts/application';