+++ src/commands/clear-pat.ts
@@ -3,7 +3,7 @@ import TYPES from './base/types';
 import { inject, injectable} from 'inversify';
 import { DELETE_PAT } from '../consts/commands';
 import { PAT_CLEAR_QUESTION, PAT_CLEARED_SUCCESS,
-	GENERIC_YES_ANSWER, GENERIC_NO_ANSWER } from '../consts/messages';
+				 GENERIC_YES_ANSWER, GENERIC_NO_ANSWER } from '../consts/constants-messages';
 import { Command } from './base/command';
 import { PersonalAccessTokenManager } from '../services/pat-manager';